fix(test-harness): close the review gaps in the gate and self-checks
- hardhat: record successful builds in the correctness table so the
  test-regression gate also covers projects whose builds are all green
- shim: record invocation arguments and require a --standard-json
  invocation in verify(), so a bare --version probe no longer counts as
  proof of compiling; moved to its own shim.rs module
- foundry: count zero-bytecode builds with zero errors as build
  failures instead of reporting them as clean
- both runners: replace the duplicated (and structurally unreachable)
  per-toolchain zero-data loops with one shared per-(project, toolchain)
  coverage tripwire in test::verify_benchmark_coverage
