+++ solx-dev/src/test/hardhat/mod.rs
@@ -37,6 +37,7 @@ pub fn test(
     })?;
 
     let mut benchmark_inputs = Vec::with_capacity(config.projects.len() * 4);
+    let mut attempted_projects = Vec::with_capacity(config.projects.len());
     let mut build_correctness_table = BTreeMap::new();
     let mut test_correctness_table = BTreeMap::new();
     let correctness_reference_compiler = config
@@ -69,7 +70,7 @@ pub fn test(
         let compiler_path = crate::utils::absolute_path(compiler.path.as_str())?;
         compiler_shims.insert(
             identifier.clone(),
-            crate::utils::CompilerInvocationShim::new(compiler_path, shim_directory.as_path())?,
+            crate::shim::CompilerInvocationShim::new(compiler_path, shim_directory.as_path())?,
         );
     }
 
@@ -84,6 +85,7 @@ pub fn test(
                         .any(|element| project_name.contains(element)))
         })
     {
+        attempted_projects.push(project_name.clone());
         let mut project_directory = crate::utils::absolute_path(projects_directory.as_path())?;
         project_directory.push(project_name.as_str());
 
@@ -289,6 +291,12 @@ pub fn test(
                 );
                 continue;
             }
+            // The correctness gate iterates this table's keys, so successes
+            // must be recorded too or the project's tests are never compared.
+            build_correctness_table
+                .entry(project_name.clone())
+                .or_insert_with(BTreeMap::new)
+                .insert(toolchain_name.clone(), 0);
             let compilation_time = build_timestamp_start.elapsed().as_millis() as u64;
             // solc toolchains compile with Hardhat's own downloaded compiler —
             // the configured path is never passed to the project, so only solx
@@ -358,26 +366,18 @@ pub fn test(
         }
     }
 
-    // A toolchain that ran but contributed nothing would be silently absent
-    // from the report, misreported as clean (see #497).
-    if !benchmark_inputs.is_empty() {
-        for ((_identifier, compiler), codegen) in config
-            .compilers
-            .iter()
-            .filter(|(_identifier, compiler)| !compiler.disabled)
-            .cartesian_product(["legacy", "viaIR"])
-        {
-            let toolchain_name = format!("{}-{codegen}", compiler.name);
-            if !benchmark_inputs
-                .iter()
-                .any(|input| input.toolchain == toolchain_name)
-            {
-                anyhow::bail!(
-                    "Harness self-check failed: toolchain {toolchain_name} is enabled but produced no benchmark data",
-                );
-            }
-        }
-    }
+    let enabled_toolchains: Vec<String> = config
+        .compilers
+        .values()
+        .filter(|compiler| !compiler.disabled)
+        .cartesian_product(["legacy", "viaIR"])
+        .map(|(compiler, codegen)| format!("{}-{codegen}", compiler.name))
+        .collect();
+    crate::test::verify_benchmark_coverage(
+        benchmark_inputs.as_slice(),
+        attempted_projects.as_slice(),
+        enabled_toolchains.as_slice(),
+    )?;
 
     let benchmark = solx_benchmark_converter::Benchmark::from_inputs(benchmark_inputs)?;
     let enabled_compiler_names: std::collections::HashSet<&str> = config
